#135857 delete without admin account

+++ app/Modules/AuthModule/Services/UserService.php
@@ -8,13 +8,18 @@ use App\Models\User;
 
 class UserService
 {
-    public function userDelete($request){
+    public function userDelete(){
         
-        // $user =User::connect(config('database.default'))->where('id',auth()->id())->first();
-        // $user->isPlayer();
-        User::connect(config('database.default'))->destroy(auth()->id());
-        $user = Auth::user();
-        $user->tokens->each->delete();
+        $user =User::connect(config('database.default'))->where('id',auth()->id())->first();
+
+        if($user->user_role_id !=2){
+
+            User::connect(config('database.default'))->destroy(auth()->id());
+            $user = Auth::user();
+            $user->tokens->each->delete();
+        }
+        
+        
 
     
     }
